requirements.txt
fastapi
gunicorn
uvicorn
opencv-python
Pillow
mediapipe